refactor(userlist): rename AddToList components to AddSpellToList and move state
